+++ CHANGELOG.md
@@ -1,21 +1,27 @@
-<a name="0.1.0"></a>
-# [0.1.0](http://test.eluying.com:91/frontend/dd-vue-component/compare/1f88fa7...v0.1.0) (2016-09-29)
+<a name="0.2.0"></a>
+# [0.2.0](http://test.eluying.com:91/frontend/dd-vue-component/compare/1f88fa7...v0.2.0) (2016-09-30)
 
 
 ### Bug Fixes
 
+* fix style rules ([8ccfcd9](http://test.eluying.com:91/frontend/dd-vue-component/commits/8ccfcd9))
+* fix style rules of pagination component ([cbff2cb](http://test.eluying.com:91/frontend/dd-vue-component/commits/cbff2cb))
 * input date can not set date ([26379a8](http://test.eluying.com:91/frontend/dd-vue-component/commits/26379a8))
+* select multiple ([3b499e8](http://test.eluying.com:91/frontend/dd-vue-component/commits/3b499e8))
 
 
 ### Features
 
 * add dd-datepicker ([3c9146e](http://test.eluying.com:91/frontend/dd-vue-component/commits/3c9146e))
+* add multiple select ([3b1e06e](http://test.eluying.com:91/frontend/dd-vue-component/commits/3b1e06e))
+* add new component dropdown ([c947fab](http://test.eluying.com:91/frontend/dd-vue-component/commits/c947fab))
 * add pagination component ([9686419](http://test.eluying.com:91/frontend/dd-vue-component/commits/9686419))
 * add select ([021dfb0](http://test.eluying.com:91/frontend/dd-vue-component/commits/021dfb0))
 * add tree ([bed9d70](http://test.eluying.com:91/frontend/dd-vue-component/commits/bed9d70))
 * feat: add component of pagination ([36240fa](http://test.eluying.com:91/frontend/dd-vue-component/commits/36240fa))
 * feat: test the example of datepicker ([61aeac0](http://test.eluying.com:91/frontend/dd-vue-component/commits/61aeac0))
 * init datepicker ([1f88fa7](http://test.eluying.com:91/frontend/dd-vue-component/commits/1f88fa7))
+* merge new feat select from dev ([d3d00a5](http://test.eluying.com:91/frontend/dd-vue-component/commits/d3d00a5))